Revert "work around Sec-Fetch-Site check"
This was based on a misunderstanding of what the problem was. I'm still
a bit unclear on how Sec-Fetch-Site interacts with redirects, but even
if I do work around that, the reciprocity to OAuth redirect *is* a
same-site redirect for me in particular.

See https://github.com/bluesky-social/atproto/issues/3362

This reverts commit 8d4046a8144a8602ce9e6d33fd17c98c8cc5a095.